+++ configure.py
@@ -937,6 +937,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/commitlog/commitlog_entry.cc',
                 'db/data_listeners.cc',
                 'db/functions/function.cc',
+                'db/hints/internal/hint_endpoint_manager.cc',
                 'db/hints/manager.cc',
                 'db/hints/resource_manager.cc',
                 'db/hints/host_filter.cc',
